Extract clipboard-adapter script
This script will deal with all the different clipboard tools that are
supported, and also allows users to provide their own versions in case
they want support for a clipboard manager that is not yet supported in
rofi-emoji.

Currently it only supports xsel/xdotool combinations, but this could be
improved on.