Allow clipboard adapter script to be replaced
In nix, this script will be replaced with a different script that wraps
the full installation env and then calls the original script. This
wrapper script will require a bash interpreter to work correctly.

By executing the script directly, it can now be replaced with a script
using a completely different interpreter on your system.

Refs #19.