Add menu system for emoji selection
Use Custom Action keybind to open menu for an entry (default Shift+Enter
AFAIK). This menu will show more information about the emoji, including
the group, subgroup, and keywords that are available for this emoji.

Right now you only get a "Back" and "Copy" action inside of the menu,
but this can be expanded freely at this point to include other useful
actions.