Release version 4.0.0
This version breaks away from Rofi 1.7 and starts to work against the as of yet
unreleased Rofi 1.8. This is because this version of Rofi contains breaking
changes to the Plugin ABI, and a lot of Rofi forks have started to appear that
are based on the 1.8 branch.

If you want to use Rofi 1.7, stick to the 3.x branch.